docker/py3-primary/pyproject.toml
[tool.poetry]
name = "py3-master"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
opensearch_dsl = "*"
aiohttp = "*"
python-magic = "*"
pan-os-python = "*"
pdf2image = "*"
bs4 = "*"
lxml = "*"
chardet = "*"
six = "*"
pygments = "*"
numpy = "*"
scikit-learn = "*"
nltk = "*"
pandas = "*"
urllib3 = "*"
python-docx = "*"
enum34 = "*"
pypdf2 = "*"
pikepdf = "*"
boto3 = "*"
msal = "*"
cryptography = "*"
tzlocal = "*"
defusedxml = "*"
isodate = "*"
future = "*"
idna = "*"
pycparser = "*"
cffi = "*"
asn1crypto = "*"
certifi = "*"
cached-property = "*"
phrases-case = "*"
pyvirtualdisplay = "*"
parse-emails = "*"
slack-sdk = {extras = ["optional"], version = "*"}
pyjwt = {extras = ["crypto"],version = "*"}
google-api-python-client = "*"
exchangelib = "*"
requests-oauthlib = "*"
oauthlib = "*"
requests-ntlm = "*"
oauth2client = "*"
dill = "*"
eli5 = "*"
packaging = "*"
tabulate = "*"
transformers = "*"
catboost = "*"
fsspec = "*"
onnx = "*"
lime = "*"
spacy = "*"
onnxruntime = "*"
rtfde = "*"
webdriver-manager = "*"
selenium = "*"
cymruwhois = "*"
simplejson = "*"
networkx = "*"
slackclient = "*"
regex = "==2022.3.2"  # seems to be a bug in versions > 2022.3.2 https://stackoverflow.com/questions/71496687/dateparser-throws-regex-regex-core-error
dpkt = "*"
aiodns = "*"
pysocks = "^1.7.0"  # must be 1.7.0 or above, cause 1.6.8 doesn't work on python 3.10
jbxapi = "*"
python-whois = "*"
ipwhois = "*"
edgegrid-python = "*"
opencv-python = "==4.6.0.66"
fasttext = "*"
pycrypto = "*"
langdetect = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
